Google Image Charts is deprecated
Hello!

Google Image Charts is deprecated and the servers are running on fumes:

> The service is running on fumes: as soon as the current serving jobs go 
down, they are unlikely to come back up. I would migrate off immediately.

https://groups.google.com/forum/#!original/google-chart-api/rZtHTyYgyXI/z67NKmx7DAAJ

We've built an alternative back in 2016 and it's now a stable, viable company that let us run a drop-in replacement and compatible API for *free*!

Enjoy!